KDTree is now a template class
