--- src/kdtree.cpp
@@ -1,172 +1,0 @@
-#include "kdtree.hpp"
-
-#include <queue>
-
-//this method tries to insert the node at this level, but if there is already
-//a node in its place, it returns a pointer to that node, else, it returns a
-//pointer to the node it tried to insert
-Node* KDTree::tryInsert(Node* finalNode, bool isXlevel, Node* toIns)
-{
-    //if this is an X level
-        //if loc.x is less than finalNode's x
-            //if there is no left child, then this IS final level
-            //else NOT
-        //else
-            //if there is no right child, then this IS final level
-            //else NOT
-
-    //if this is a Y level
-        //if loc.y is less than finalNode's y
-            //if there is no left child, then this IS final level
-            //else NOT
-        //else
-            //if there is no right child, then this IS final level
-            //else NOT
-    
-    //if this is an x level, then we compare x values, else y values
-    float nodeCompValue = isXlevel ? finalNode->loc.x() : finalNode->loc.y();
-    float insCompValue  = isXlevel ?     toIns->loc.x() : toIns->loc.y()    ;
-
-    if (insCompValue < nodeCompValue)
-    { //work on the left side of the tree
-        if (finalNode->left == NULL)
-        { //if the new node can be inserted, do so
-            finalNode->left = toIns;
-        }
-        return finalNode->left;
-    }
-    else
-    { //work on the right side of the tree
-        if (finalNode->right == NULL)
-        {
-            finalNode->right = toIns;
-        }
-        return finalNode->right;
-    }
-}
-
-void KDTree::insert(Point loc, size_t idx)
-{
-    Node* newNode = new Node(loc, idx);
-
-    if (root == NULL)
-    {
-        root = newNode;
-        return;
-    }
-    
-
-    Node* finalNode = root;
-    Node* nextNode = NULL;
-    bool isXlevel = true;
-    //tryInsert() will return a pointer to the node at the next level. If there
-    //wasn't a node there before, then it will insert newNode and return a
-    //pointer to it, since it is the node at the next level. This loop will run
-    //until that happens
-    while ((nextNode=tryInsert(finalNode, isXlevel, newNode)) != newNode)
-    {
-        finalNode = nextNode;
-        isXlevel = !isXlevel;
-    }
-    //loop invariant: finalNode points to the node that the new node will be
-    //inserted under (unless the next iteration finds another level)
-
-    allNodes.push_back(newNode);
-}
-
-struct rangeSearchStruct
-{
-    Node *node;
-    bool isXlevel;
-};
-std::vector<size_t> KDTree::rangeSearch(Point loc, float radius)
-{
-    std::vector<size_t> foundIdxes;
-
-    std::queue<rangeSearchStruct> toSearch;
-    //insert the root node, and mark it as a level sorted by X
-    toSearch.push({.node=root, .isXlevel=true});
-    while (toSearch.size() > 0)
-    {
-        //extract a node to evaluate
-        Node *v       = toSearch.front().node;
-        bool isXlevel = toSearch.front().isXlevel;
-        toSearch.pop();
-
-        //if the node doesn't exist, don't evaluate it
-        if (v == NULL) {
-            continue;
-        }
-
-        //figure out which values to compate
-        float nodeCompValue, otherNodeCompValue;
-        float queryCompValue, otherQueryCompValue;
-        if (isXlevel)
-        { //if this is an X level, evaluate the x values
-            nodeCompValue = v->loc.x();
-            otherNodeCompValue = v->loc.y();
-            queryCompValue = loc.x();
-            otherQueryCompValue = loc.y();
-        }
-        else
-        {
-            nodeCompValue = v->loc.y();
-            otherNodeCompValue = v->loc.x();
-            queryCompValue = loc.y();
-            otherQueryCompValue = loc.x();
-        }
-
-        //based on the relevant dimension, decide what to do
-        if (nodeCompValue < (queryCompValue - radius))
-        { //the node cannot be in range, but some of its right children could be
-            toSearch.push({.node=v->right, .isXlevel=!isXlevel});
-        }
-        else if (
-            (queryCompValue - radius) <= nodeCompValue &&
-            nodeCompValue < (queryCompValue + radius))
-        {
-            //if the relevant dimension is in range, its irrelevant dimension could be, too
-            if
-            (
-                (otherQueryCompValue - radius) <= otherNodeCompValue &&
-                 otherNodeCompValue < (otherQueryCompValue + radius)
-            )
-            {
-                foundIdxes.push_back(v->idx);
-            }
-            //descendants in both trees should be searched
-            toSearch.push({.node=v->left , .isXlevel=!isXlevel});
-            toSearch.push({.node=v->right, .isXlevel=!isXlevel});
-        }
-        else
-        {
-            toSearch.push({.node=v->left, .isXlevel=!isXlevel});
-        }
-    }
-    
-    return foundIdxes;
-}
-
-std::vector<size_t> KDTree::breadthFirstSearch() {
-    std::vector<size_t> discovered;
-    
-    std::queue<Node*> q;
-
-    discovered.push_back(root->idx);
-    q.push(root);
-    while (q.size() > 0)
-    {
-        Node* v = q.front();
-        q.pop();
-        if (v->left != NULL) {
-            discovered.push_back(v->left->idx);
-            q.push(v->left);
-        }
-        if (v->right != NULL) {
-            discovered.push_back(v->right->idx);
-            q.push(v->right);
-        }
-    }
-
-    return discovered;
-}
